# packages/contracts/openapi.yaml
components:
  schemas:
    AnswerResponse:
      properties:
        answer:
          title: Answer
          type: string
        answer_status:
          enum:
          - answered
          - insufficient_evidence
          title: Answer Status
          type: string
        cards:
          items:
            $ref: '#/components/schemas/CardResult'
          title: Cards
          type: array
        claims:
          items:
            $ref: '#/components/schemas/AtomicClaim'
          title: Claims
          type: array
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceResult'
          title: Evidence
          type: array
        profile:
          enum:
          - card_page_section_benefit
          - parent_child_bundle
          title: Profile
          type: string
        query_type:
          enum:
          - proper_noun
          - numeric_condition
          - semantic
          title: Query Type
          type: string
        recommendations:
          items:
            $ref: '#/components/schemas/Recommendation'
          title: Recommendations
          type: array
        release_id:
          title: Release Id
          type: string
        status:
          const: completed
          title: Status
          type: string
        usage:
          $ref: '#/components/schemas/AnswerUsage'
      required:
      - status
      - answer_status
      - release_id
      - profile
      - query_type
      - cards
      - answer
      - recommendations
      - claims
      - evidence
      - usage
      title: AnswerResponse
      type: object
    AnswerUsage:
      properties:
        answer:
          additionalProperties: true
          title: Answer
          type: object
        embedding:
          additionalProperties: true
          title: Embedding
          type: object
      required:
      - embedding
      - answer
      title: AnswerUsage
      type: object
    AtomicClaim:
      additionalProperties: false
      properties:
        card_key:
          maxLength: 64
          minLength: 1
          title: Card Key
          type: string
        citations:
          items:
            maxLength: 64
            minLength: 1
            type: string
          maxItems: 2
          minItems: 1
          title: Citations
          type: array
        conditions:
          items:
            maxLength: 60
            minLength: 1
            type: string
          maxItems: 2
          title: Conditions
          type: array
        text:
          maxLength: 120
          minLength: 1
          title: Text
          type: string
        unit:
          anyOf:
          - maxLength: 20
            type: string
          - type: 'null'
          title: Unit
        value:
          anyOf:
          - maxLength: 40
            type: string
          - type: number
          - type: integer
          - type: 'null'
          title: Value
      required:
      - card_key
      - text
      - citations
      title: AtomicClaim
      type: object
    BrowserSessionRequest:
      additionalProperties: false
      properties: {}
      title: BrowserSessionRequest
      type: object
    BrowserSessionResponse:
      properties:
        status:
          const: ready
          title: Status
          type: string
      required:
      - status
      title: BrowserSessionResponse
      type: object
    CardResult:
      properties:
        card_key:
          title: Card Key
          type: string
        card_name:
          title: Card Name
          type: string
        evidence_count:
          title: Evidence Count
          type: integer
        issuer:
          title: Issuer
          type: string
        rank:
          title: Rank
          type: integer
        score:
          title: Score
          type: number
      required:
      - card_key
      - card_name
      - issuer
      - score
      - rank
      - evidence_count
      title: CardResult
      type: object
    ChatMessage:
      properties:
        answer:
          anyOf:
          - $ref: '#/components/schemas/AnswerResponse'
          - type: 'null'
        client_request_id:
          title: Client Request Id
          type: string
        content:
          title: Content
          type: string
        created_at:
          title: Created At
          type: string
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorResponse'
          - type: 'null'
        id:
          title: Id
          type: string
        rewrite_usage:
          anyOf:
          - $ref: '#/components/schemas/ChatRewriteUsage'
          - type: 'null'
        role:
          enum:
          - user
          - assistant
          title: Role
          type: string
        seq:
          title: Seq
          type: integer
        status:
          enum:
          - completed
          - pending
          - failed
          title: Status
          type: string
        turn_id:
          title: Turn Id
          type: string
      required:
      - id
      - turn_id
      - client_request_id
      - seq
      - role
      - content
      - status
      - created_at
      title: ChatMessage
      type: object
    ChatRequest:
      additionalProperties: false
      properties:
        client_request_id:
          format: uuid
          title: Client Request Id
          type: string
        profile:
          anyOf:
          - enum:
            - card_page_section_benefit
            - parent_child_bundle
            type: string
          - type: 'null'
          title: Profile
        query:
          maxLength: 500
          minLength: 1
          title: Query
          type: string
        retry_failed:
          default: false
          title: Retry Failed
          type: boolean
        top_k:
          default: 3
          enum:
          - 1
          - 3
          - 5
          title: Top K
          type: integer
      required:
      - query
      - client_request_id
      title: ChatRequest
      type: object
    ChatRewriteUsage:
      properties:
        latency_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Latency Ms
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        provider_called:
          title: Provider Called
          type: boolean
        usage:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Usage
      required:
      - provider_called
      title: ChatRewriteUsage
      type: object
    Conversation:
      properties:
        created_at:
          title: Created At
          type: string
        id:
          title: Id
          type: string
        title:
          title: Title
          type: string
        updated_at:
          title: Updated At
          type: string
      required:
      - id
      - title
      - created_at
      - updated_at
      title: Conversation
      type: object
    ConversationPage:
      properties:
        conversations:
          items:
            $ref: '#/components/schemas/Conversation'
          title: Conversations
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      required:
      - conversations
      - next_cursor
      title: ConversationPage
      type: object
    CreateConversationRequest:
      additionalProperties: false
      properties:
        client_conversation_id:
          format: uuid
          title: Client Conversation Id
          type: string
      required:
      - client_conversation_id
      title: CreateConversationRequest
      type: object
    ErrorResponse:
      properties:
        code:
          title: Code
          type: string
        message:
          title: Message
          type: string
        request_id:
          title: Request Id
          type: string
        retryable:
          title: Retryable
          type: boolean
      required:
      - code
      - message
      - retryable
      - request_id
      title: ErrorResponse
      type: object
    EvidenceResult:
      properties:
        card_key:
          title: Card Key
          type: string
        card_name:
          title: Card Name
          type: string
        chunk_id:
          title: Chunk Id
          type: string
        issuer:
          title: Issuer
          type: string
        level:
          title: Level
          type: string
        page_num:
          title: Page Num
          type: integer
        rank:
          title: Rank
          type: integer
        score:
          title: Score
          type: number
        section:
          anyOf:
          - type: string
          - type: 'null'
          title: Section
        text:
          title: Text
          type: string
      required:
      - rank
      - card_key
      - card_name
      - issuer
      - chunk_id
      - page_num
      - text
      - section
      - level
      - score
      title: EvidenceResult
      type: object
    LiveResponse:
      properties:
        status:
          const: live
          title: Status
          type: string
      required:
      - status
      title: LiveResponse
      type: object
    MessagesPage:
      properties:
        has_pending:
          title: Has Pending
          type: boolean
        messages:
          items:
            $ref: '#/components/schemas/ChatMessage'
          title: Messages
          type: array
        next_before_seq:
          anyOf:
          - type: integer
          - type: 'null'
          title: Next Before Seq
      required:
      - messages
      - next_before_seq
      - has_pending
      title: MessagesPage
      type: object
    NotReadyResponse:
      properties:
        reason:
          title: Reason
          type: string
        status:
          const: not_ready
          title: Status
          type: string
      required:
      - status
      - reason
      title: NotReadyResponse
      type: object
    QueryRequest:
      additionalProperties: false
      properties:
        profile:
          anyOf:
          - enum:
            - card_page_section_benefit
            - parent_child_bundle
            type: string
          - type: 'null'
          title: Profile
        query:
          maxLength: 500
          minLength: 1
          title: Query
          type: string
        top_k:
          default: 3
          enum:
          - 1
          - 3
          - 5
          title: Top K
          type: integer
      required:
      - query
      title: QueryRequest
      type: object
    ReadyResponse:
      properties:
        chunk_count:
          title: Chunk Count
          type: integer
        document_count:
          title: Document Count
          type: integer
        profile:
          enum:
          - card_page_section_benefit
          - parent_child_bundle
          title: Profile
          type: string
        release_id:
          title: Release Id
          type: string
        status:
          const: ready
          title: Status
          type: string
      required:
      - status
      - release_id
      - profile
      - document_count
      - chunk_count
      title: ReadyResponse
      type: object
    Recommendation:
      additionalProperties: false
      properties:
        card_key:
          maxLength: 64
          minLength: 1
          title: Card Key
          type: string
        citations:
          items:
            maxLength: 64
            minLength: 1
            type: string
          maxItems: 2
          minItems: 1
          title: Citations
          type: array
        reason:
          maxLength: 100
          minLength: 1
          title: Reason
          type: string
      required:
      - card_key
      - reason
      - citations
      title: Recommendation
      type: object
    SearchResponse:
      properties:
        cards:
          items:
            $ref: '#/components/schemas/CardResult'
          title: Cards
          type: array
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceResult'
          title: Evidence
          type: array
        profile:
          enum:
          - card_page_section_benefit
          - parent_child_bundle
          title: Profile
          type: string
        query_type:
          enum:
          - proper_noun
          - numeric_condition
          - semantic
          title: Query Type
          type: string
        release_id:
          title: Release Id
          type: string
        status:
          const: completed
          title: Status
          type: string
        usage:
          $ref: '#/components/schemas/SearchUsage'
      required:
      - status
      - release_id
      - profile
      - query_type
      - cards
      - evidence
      - usage
      title: SearchResponse
      type: object
    SearchUsage:
      properties:
        embedding:
          additionalProperties: true
          title: Embedding
          type: object
      required:
      - embedding
      title: SearchUsage
      type: object
    TurnResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/ChatMessage'
          title: Messages
          type: array
        turn_id:
          title: Turn Id
          type: string
      required:
      - turn_id
      - messages
      title: TurnResponse
      type: object
info:
  title: PickCardU RAG API
  version: 0.1.0
openapi: 3.1.0
paths:
  /v1/answer:
    post:
      operationId: answer_v1_answer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
          description: Successful Response
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Answer
  /v1/browser-session:
    post:
      operationId: initialize_session_v1_browser_session_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserSessionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserSessionResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Initialize Session
  /v1/conversations:
    get:
      operationId: conversations_v1_conversations_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 40
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: cursor
        required: false
        schema:
          anyOf:
          - maxLength: 1024
            type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationPage'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Conversations
    post:
      operationId: create_conversation_v1_conversations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
          description: OK
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Create Conversation
  /v1/conversations/{conversation_id}/messages:
    get:
      operationId: messages_v1_conversations__conversation_id__messages_get
      parameters:
      - in: path
        name: conversation_id
        required: true
        schema:
          format: uuid
          title: Conversation Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: before_seq
        required: false
        schema:
          anyOf:
          - minimum: 1
            type: integer
          - type: 'null'
          title: Before Seq
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesPage'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Messages
    post:
      operationId: send_message_v1_conversations__conversation_id__messages_post
      parameters:
      - in: path
        name: conversation_id
        required: true
        schema:
          format: uuid
          title: Conversation Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Send Message
  /v1/health/live:
    get:
      operationId: live_v1_health_live_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveResponse'
          description: Successful Response
      summary: Live
  /v1/health/ready:
    get:
      operationId: ready_v1_health_ready_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
          description: Successful Response
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotReadyResponse'
          description: Service Unavailable
      summary: Ready
  /v1/search:
    post:
      operationId: search_v1_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
          description: Successful Response
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
      summary: Search